requirements.txt
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
Deprecated==1.2.14
Flask==2.3.2
googleapis-common-protos==1.56.2
grpcio==1.57.0
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
opentelemetry-api==1.19.0
opentelemetry-distro==0.40b0
opentelemetry-exporter-jaeger==1.19.0
opentelemetry-exporter-jaeger-proto-grpc==1.19.0
opentelemetry-exporter-jaeger-thrift==1.19.0
opentelemetry-instrumentation==0.40b0
opentelemetry-instrumentation-aws-lambda==0.40b0
opentelemetry-instrumentation-dbapi==0.40b0
opentelemetry-instrumentation-flask==0.40b0
opentelemetry-instrumentation-jinja2==0.40b0
opentelemetry-instrumentation-logging==0.40b0
opentelemetry-instrumentation-sqlite3==0.40b0
opentelemetry-instrumentation-urllib==0.40b0
opentelemetry-instrumentation-wsgi==0.40b0
opentelemetry-propagator-aws-xray==1.0.1
opentelemetry-sdk==1.19.0
opentelemetry-semantic-conventions==0.40b0
opentelemetry-util-http==0.40b0
packaging==23.1
protobuf==3.20.3
requests==2.31.0
six==1.16.0
thrift==0.16.0
typing_extensions==4.7.1
urllib3==2.0.4
Werkzeug==2.3.7
wrapt==1.15.0
zipp==3.16.2
